NCS-87 Moving shareholder files into subfolders to follow the oracle-query-api folder structure
