+++ src/test/java/uk/gov/ch/service/shareholder/impl/ShareholderServiceImplTest.java
@@ -1,18 +1,17 @@
-package uk.gov.ch.service.impl;
+package uk.gov.ch.service.shareholder.impl;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import uk.gov.ch.dao.ShareholderDao;
-import uk.gov.ch.model.Shareholder;
+import uk.gov.ch.repository.shareholder.ShareholderRepository;
+import uk.gov.ch.model.shareholder.Shareholder;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -23,14 +22,14 @@ class ShareholderServiceImplTest {
     private static final String COMPANY_NUMBER = "123456789";
 
     @Mock
-    private ShareholderDao shareholderDao;
+    private ShareholderRepository shareholderRepository;
 
     @InjectMocks
     private ShareholderServiceImpl shareholderService;
 
     @Test
     void getShareholderCount() {
-        when(shareholderDao.getShareholderCount(any())).thenReturn(3);
+        when(shareholderRepository.getShareholderCount(any())).thenReturn(3);
 
         int result = shareholderService.getShareholderCount(COMPANY_NUMBER);
 
@@ -42,7 +41,7 @@ class ShareholderServiceImplTest {
         List<Shareholder> expectedList = new ArrayList<Shareholder>();
         expectedList.add(new Shareholder());
 
-        when(shareholderDao.getShareholders(any())).thenReturn(expectedList);
+        when(shareholderRepository.getShareholders(any())).thenReturn(expectedList);
 
         List<Shareholder> result = shareholderService.getShareholders(COMPANY_NUMBER);
